Set up separate toolchain/codebase for the JS UI
* set up a separate NPM project;
* split the monolithic ui.js into separate files (one per component);
* split out the sample program into its own file;
* move static files to the JS project;
* add .gitignore configuration.
