+++ .gitignore
@@ -117,5 +117,41 @@ edumips64.iml
 /bazel-*
 .project/.blaze/
 
+#################################################################################
+## Misc
+
 # Mac OS X
 .DS_Store
+
+#################################################################################
+## Node
+
+# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+lerna-debug.log*
+
+# Diagnostic reports (https://nodejs.org/api/report.html)
+report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
+
+# Runtime data
+pids
+*.pid
+*.seed
+*.pid.lock
+
+# Dependency directories
+node_modules/
+jspm_packages/
+
+# Optional npm cache directory
+.npm
+
+# Optional eslint cache
+.eslintcache
+
+# Output directory
+dist/
