+++ src/webapp/src/components/Memory.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Memory = (props) => {
+    return (
+        <div id="memory">
+            <textarea readOnly value={props.memory} />
+        </div>
+    );
+}
+
+export default Memory;
