+++ src/webapp/src/static/favicon.png
