+++ src/webapp/src/static/index.html
@@ -10,12 +10,7 @@
       <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
       <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
 
-      <!-- JS libraries: Require, React, Babel (for JSX). -->
-      <script src="https://unpkg.com/react@16/umd/react.development.js" crossorigin></script>
-      <script src="https://unpkg.com/react-dom@16/umd/react-dom.development.js" crossorigin></script>
-      <script src="https://unpkg.com/@babel/standalone@7.9.6/babel.min.js"></script>
-
-      <!-- JS initialization code -->
+      <!-- Materialize JS initialization code -->
       <script type="text/javascript" language="javascript">
         // Initialization code for Materialize.
         document.addEventListener('DOMContentLoaded', () => {
@@ -24,8 +19,8 @@
         });
       </script>
 
-      <!-- Import the React UI code. This is pre-processed in-browser (JSX -> JS) using Babel. -->
-      <script type="text/babel" src="ui.js" ></script>
+      <!-- Import the React UI code. -->
+      <script type="text/javascript" src="ui.js"></script>
 
     </script>
   </head>
