+++ src/webapp/webpack.config.js
@@ -1,0 +1,14 @@
+const path = require("path");
+
+module.exports = {
+    entry: "./src/index.js",
+    output: {
+        path: path.resolve(__dirname, "dist"),
+        filename: "ui.js"
+    },
+    module: {
+        rules: [
+            {test: /\.js$/, exclude: /node_modules/, loader: "babel-loader"}
+        ]
+    }
+}
